Implement the `INPUT` command for user interaction
*   Add parsing and handling for `INPUT` statements, supporting optional prompts and variable assignment.
*   Refactor `animate_input_cursor` and `handle_input_key` into `App` methods and integrate cursor animation into the render loop.
*   Update `draw_glyph` to support independent X and Y scaling for more flexible rendering, including the input cursor.
*   Move `ClsColor` enum definition to `src/colors.rs` for better organization.

+++ src/glyphs.rs
@@ -302,7 +302,7 @@ pub fn get_glyph(ch: char) -> [i8; GLYPH_PIXELS] {
 }
 
 #[inline]
-pub fn glyph_pixels(id: u16) -> [i8; GLYPH_PIXELS] { GLYPHS[id] }
+pub fn glyph_pixels(id: usize) -> [i8; GLYPH_PIXELS] { GLYPHS[id] }
 
 #[inline]
 pub fn get_glyph_ref(ch: char) -> &'static [i8; GLYPH_PIXELS] {
